+++ inst/example/LDS_C10.R
@@ -27,14 +27,13 @@ LDS_C10_RESPDIS
 
 
 # Page 536, Question 3 (Large Data Set ‘RISKFACT’)
-head(RISKFACT_raw <- read.csv('data/LDS_C10_RISKFACT.csv'))
-dim(RISKFACT_raw)
-names(RISKFACT_raw)[2:6] = c('O2', 'SBP', 'TChol', 'HDLChol', 'TriG')
-
-scatterplotMatrix(~ O2 + SBP + TChol + HDLChol + TriG, data = RISKFACT_raw, smooth = FALSE, regLine = FALSE)
+RISKFACT0 = read.csv(system.file('extdata', 'LDS_C10_RISKFACT_corrected.csv', 
+                                package = 'DanielBiostatistics10th'))
+head(RISKFACT0)
+car::scatterplotMatrix(~ O2 + SBP + TChol + HDLChol + TriG, data = RISKFACT0, smooth = FALSE, regLine = FALSE)
 # There are obvious typos in SBP, the corresponding records should be removed
-dim(RISKFACT <- RISKFACT_raw[-c(which.max(RISKFACT_raw$SBP), which.min(RISKFACT_raw$SBP)), , drop = FALSE])
-scatterplotMatrix(~ O2 + SBP + TChol + HDLChol + TriG, data = RISKFACT, smooth = FALSE, regLine = FALSE)
+dim(RISKFACT <- RISKFACT0[-c(which.max(RISKFACT0$SBP), which.min(RISKFACT0$SBP)), , drop = FALSE])
+car::scatterplotMatrix(~ O2 + SBP + TChol + HDLChol + TriG, data = RISKFACT, smooth = FALSE, regLine = FALSE)
 
 summary(mod_RISKFACT <- lm(O2 ~ SBP + TChol + HDLChol + TriG, data = RISKFACT))
 # 84% of total variance is explained by the multivariable linear regression model